requirements.txt
pywebview==3.4